# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Jersey OS API
  version: 1.0.0
  description: Platform and product catalog HTTP contract for Jersey OS.
servers:
  - url: /api/v1
paths:
...
  /catalog/products:
    get:
      operationId: searchProducts
      tags: [Catalog]
      summary: Search products
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/ProductStatus' }
        - in: query
          name: teamId
          schema: { type: string, format: uuid }
        - in: query
          name: seasonId
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedProductsResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      operationId: createProduct
      tags: [Catalog]
      summary: Create a draft product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Validation or business rule failure
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /catalog/products/{productId}:
    get:
      operationId: getProduct
      tags: [Catalog]
      summary: Get product detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
    put:
      operationId: updateProduct
      tags: [Catalog]
      summary: Update product details and taxonomy links
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
  /catalog/products/{productId}/activate:
    post:
      operationId: activateProduct
      tags: [Catalog]
      summary: Activate a product
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Activation rules failed
        '404':
          description: Not found
  /catalog/products/{productId}/archive:
    post:
      operationId: archiveProduct
      tags: [Catalog]
      summary: Archive a product
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
  /catalog/products/{productId}/variants:
    put:
      operationId: upsertVariant
      tags: [Catalog]
      summary: Create or update a product variant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertVariantRequest'
      responses:
        '200':
          description: Product with variants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
  /catalog/products/{productId}/variants/{variantId}:
    delete:
      operationId: removeVariant
      tags: [Catalog]
      summary: Remove a product variant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/VariantId'
      responses:
        '200':
          description: Product without the variant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
  /catalog/products/{productId}/images:
    post:
      operationId: attachProductImage
      tags: [Catalog]
      summary: Upload and attach a product image
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                altText:
                  type: string
                sortOrder:
                  type: integer
      responses:
        '200':
          description: Product with images
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
  /catalog/products/{productId}/images/order:
    put:
      operationId: reorderProductImages
      tags: [Catalog]
      summary: Reorder product images
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderImagesRequest'
      responses:
        '200':
          description: Reordered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
  /catalog/products/{productId}/images/{imageId}:
    delete:
      operationId: removeProductImage
      tags: [Catalog]
      summary: Remove a product image
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - in: path
          name: imageId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Image removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not found
  /catalog/teams:
    get:
      operationId: listTeams
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxonomyItemResponse'
    post:
      operationId: createTeam
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxonomyItemRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyItemResponse'
  /catalog/teams/{id}:
    put:
      operationId: updateTeam
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaxonomyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaxonomyItemRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyItemResponse'
        '404':
          description: Not found
  /catalog/seasons:
    get:
      operationId: listSeasons
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Seasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxonomyItemResponse'
    post:
      operationId: createSeason
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxonomyItemRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyItemResponse'
  /catalog/seasons/{id}:
    put:
      operationId: updateSeason
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaxonomyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaxonomyItemRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyItemResponse'
        '404':
          description: Not found
  /catalog/categories:
    get:
      operationId: listCategories
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxonomyItemResponse'
    post:
      operationId: createCategory
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxonomyItemRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyItemResponse'
  /catalog/categories/{id}:
    put:
      operationId: updateCategory
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaxonomyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaxonomyItemRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyItemResponse'
        '404':
          description: Not found
  /catalog/tags:
    get:
      operationId: listTags
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxonomyItemResponse'
    post:
      operationId: createTag
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxonomyItemRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyItemResponse'
  /catalog/tags/{id}:
    put:
      operationId: updateTag
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaxonomyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaxonomyItemRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyItemResponse'
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductId:
      in: path
      name: productId
      required: true
      schema: { type: string, format: uuid }
    VariantId:
      in: path
      name: variantId
      required: true
      schema: { type: string, format: uuid }
    TaxonomyId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    ProductStatus:
      type: string
      enum: [Draft, Active, Archived]
    TaxonomyItemResponse:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        parentId: { type: string, format: uuid, nullable: true }
    CreateTaxonomyItemRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        parentId: { type: string, format: uuid, nullable: true }
    UpdateTaxonomyItemRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        parentId: { type: string, format: uuid, nullable: true }
    InventoryResponse:
      type: object
      required: [variantId, onHand, reserved, rowVersion]
      properties:
        variantId: { type: string, format: uuid }
        onHand: { type: integer }
        reserved: { type: integer }
        rowVersion: { type: string }
    ProductVariantResponse:
      type: object
      required: [id, sku, size, sortOrder, inventory]
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        size: { type: string }
        sortOrder: { type: integer }
        inventory:
          $ref: '#/components/schemas/InventoryResponse'
    ProductImageResponse:
      type: object
      required: [id, objectKey, url, contentType, sortOrder]
      properties:
        id: { type: string, format: uuid }
        objectKey: { type: string }
        url: { type: string }
        contentType: { type: string }
        altText: { type: string, nullable: true }
        sortOrder: { type: integer }
    ProductResponse:
      type: object
      required: [id, name, slug, status, categoryIds, tagIds, variants, images]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        styleCode: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ProductStatus' }
        teamId: { type: string, format: uuid, nullable: true }
        seasonId: { type: string, format: uuid, nullable: true }
        categoryIds:
          type: array
          items: { type: string, format: uuid }
        tagIds:
          type: array
          items: { type: string, format: uuid }
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariantResponse'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImageResponse'
    ProductSummaryResponse:
      type: object
      required: [id, name, slug, status, variantCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        styleCode: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ProductStatus' }
        teamId: { type: string, format: uuid, nullable: true }
        seasonId: { type: string, format: uuid, nullable: true }
        variantCount: { type: integer }
    PagedProductsResponse:
      type: object
      required: [items, page, pageSize, totalCount]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummaryResponse'
        page: { type: integer }
        pageSize: { type: integer }
        totalCount: { type: integer }
    CreateProductRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        styleCode: { type: string, nullable: true }
        teamId: { type: string, format: uuid, nullable: true }
        seasonId: { type: string, format: uuid, nullable: true }
        categoryIds:
          type: array
          items: { type: string, format: uuid }
        tagIds:
          type: array
          items: { type: string, format: uuid }
    UpdateProductRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        styleCode: { type: string, nullable: true }
        teamId: { type: string, format: uuid, nullable: true }
        seasonId: { type: string, format: uuid, nullable: true }
        categoryIds:
          type: array
          items: { type: string, format: uuid }
        tagIds:
          type: array
          items: { type: string, format: uuid }
    UpsertVariantRequest:
      type: object
      required: [sku, size, sortOrder]
      properties:
        id: { type: string, format: uuid, nullable: true }
        sku: { type: string }
        size: { type: string }
        sortOrder: { type: integer }
    ImageSortOrder:
      type: object
      required: [imageId, sortOrder]
      properties:
        imageId: { type: string, format: uuid }
        sortOrder: { type: integer }
    ReorderImagesRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ImageSortOrder'
...